+++ example/gen.go
@@ -1,2 +1,4 @@
 //go:generate go run github.com/jmattheis/goverter/cmd/goverter gen ./...
+//go:generate go run github.com/jmattheis/goverter/cmd/goverter gen -cwd ./wrap-errors-using ./
+//go:generate go run github.com/jmattheis/goverter/cmd/goverter gen -cwd ./protobuf ./
 package example